# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vendor Control Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Plan = (typeof Plan)[keyof typeof Plan];

export const Plan = {
  starter: "starter",
  pro: "pro",
} as const;

export interface Vendor {
  id: string;
  name: string;
  phoneNumber: string;
  plan: Plan;
  botEnabled: boolean;
  bankName?: string | null;
  bankAccountNumber?: string | null;
  bankAccountHolder?: string | null;
  currency?: string;
  welcomeMessage?: string | null;
  createdAt: string;
}

export interface VendorStats {
  totalOrders: number;
  pendingOrders: number;
  confirmedOrders: number;
  paidOrders: number;
  revenue: number;
  menuItems: number;
  customers: number;
  openConversations: number;
}

export type VendorDetail = Vendor & {
  stats: VendorStats;
};

export interface CreateVendorBody {
  /** @minLength 1 */
  name: string;
  /** @minLength 4 */
  phoneNumber: string;
  plan: Plan;
  bankName?: string;
  bankAccountNumber?: string;
  bankAccountHolder?: string;
  currency?: string;
  welcomeMessage?: string;
}

export interface UpdateVendorBody {
  name?: string;
  phoneNumber?: string;
  plan?: Plan;
  botEnabled?: boolean;
  bankName?: string;
  bankAccountNumber?: string;
  bankAccountHolder?: string;
  currency?: string;
  welcomeMessage?: string;
}

export interface MenuItem {
  id: string;
  vendorId: string;
  name: string;
  description?: string | null;
  price: number;
  category?: string | null;
  available: boolean;
  createdAt: string;
}

export interface CreateMenuItemBody {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minimum 0 */
  price: number;
  category?: string;
  available?: boolean;
}

export interface UpdateMenuItemBody {
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  category?: string;
  available?: boolean;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  paid: "paid",
  rejected: "rejected",
  completed: "completed",
} as const;

export interface OrderItem {
  name: string;
  /** @minimum 1 */
  quantity: number;
  unitPrice: number;
}

export interface Order {
  id: string;
  vendorId: string;
  vendorName?: string | null;
  customerPhone: string;
  customerName: string;
  status: OrderStatus;
  total: number;
  currency?: string;
  items: OrderItem[];
  notes?: string | null;
  createdAt: string;
}

export type OrderDetail = Order;

export interface UpdateOrderBody {
  status: OrderStatus;
  note?: string;
}

export type ConversationStatus =
  (typeof ConversationStatus)[keyof typeof ConversationStatus];

export const ConversationStatus = {
  bot: "bot",
  human: "human",
  closed: "closed",
} as const;

export interface Conversation {
  id: string;
  vendorId: string;
  vendorName?: string | null;
  customerPhone: string;
  customerName: string;
  status: ConversationStatus;
  lastMessagePreview: string;
  lastMessageAt: string;
  unreadCount: number;
  createdAt: string;
}

export type MessageDirection =
  (typeof MessageDirection)[keyof typeof MessageDirection];

export const MessageDirection = {
  in: "in",
  out: "out",
} as const;

export type MessageSender = (typeof MessageSender)[keyof typeof MessageSender];

export const MessageSender = {
  customer: "customer",
  bot: "bot",
  vendor: "vendor",
  system: "system",
} as const;

export interface Message {
  id: string;
  conversationId: string;
  direction: MessageDirection;
  sender: MessageSender;
  body: string;
  createdAt: string;
}

export type ConversationDetail = Conversation & {
  messages: Message[];
};

export interface UpdateConversationBody {
  status?: ConversationStatus;
  markRead?: boolean;
}

export type SendMessageBodySender =
  (typeof SendMessageBodySender)[keyof typeof SendMessageBodySender];

export const SendMessageBodySender = {
  vendor: "vendor",
  system: "system",
} as const;

export interface SendMessageBody {
  /** @minLength 1 */
  body: string;
  sender?: SendMessageBodySender;
}

export interface Customer {
  id: string;
  vendorId: string;
  phone: string;
  name: string;
  notes?: string | null;
  totalOrders: number;
  totalSpent: number;
  lastSeenAt?: string | null;
  createdAt: string;
}

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod];

export const PaymentMethod = {
  bank_transfer: "bank_transfer",
  payment_link: "payment_link",
  cash: "cash",
} as const;

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  failed: "failed",
} as const;

export interface Payment {
  id: string;
  vendorId: string;
  orderId: string;
  customerName?: string;
  amount: number;
  currency?: string;
  method: PaymentMethod;
  status: PaymentStatus;
  reference?: string | null;
  createdAt: string;
}

export type DashboardSummaryOrdersByStatusItem = {
  status: OrderStatus;
  count: number;
};

export type DashboardSummaryRevenueByVendorItem = {
  vendorId: string;
  vendorName: string;
  revenue: number;
};

export interface DashboardSummary {
  totalVendors: number;
  starterVendors: number;
  proVendors: number;
  totalOrders: number;
  pendingOrders: number;
  revenue: number;
  openConversations: number;
  messagesToday: number;
  ordersByStatus: DashboardSummaryOrdersByStatusItem[];
  revenueByVendor: DashboardSummaryRevenueByVendorItem[];
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  order: "order",
  message: "message",
  payment: "payment",
  vendor: "vendor",
  handover: "handover",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  vendorName: string;
  createdAt: string;
}

export type VendorAnalyticsDailyOrdersItem = {
  date: string;
  count: number;
};

export type VendorAnalyticsDailyRevenueItem = {
  date: string;
  amount: number;
};

export type VendorAnalyticsTopItemsItem = {
  name: string;
  quantity: number;
  revenue: number;
};

export interface VendorAnalytics {
  dailyOrders: VendorAnalyticsDailyOrdersItem[];
  dailyRevenue: VendorAnalyticsDailyRevenueItem[];
  topItems: VendorAnalyticsTopItemsItem[];
  repeatCustomerRate: number;
}

export interface WhatsAppWebhookBody {
  /** Customer phone number */
  from: string;
  /** Vendor WhatsApp number */
  to: string;
  body: string;
  profileName?: string;
}

export interface SimulateIncomingBody {
  vendorId: string;
  customerPhone: string;
  customerName?: string;
  body: string;
}

export interface WebhookResponse {
  ok: boolean;
  botReply?: string | null;
  conversationId?: string | null;
}

export type ListVendorOrdersParams = {
  status?: ListVendorOrdersStatus;
};

export type ListVendorOrdersStatus =
  (typeof ListVendorOrdersStatus)[keyof typeof ListVendorOrdersStatus];

export const ListVendorOrdersStatus = {
  pending: "pending",
  confirmed: "confirmed",
  paid: "paid",
  rejected: "rejected",
  completed: "completed",
} as const;