# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vendor Control Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  vendorId: string;
  vendorName?: string | null;
  customerPhone: string;
  customerName: string;
  status: OrderStatus;
  total: number;
  currency?: string;
  items: OrderItem[];
  notes?: string | null;
  createdAt: Date;
}