# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vendor Control Panel API
 * OpenAPI spec version: 0.1.0
 */

export type Plan = (typeof Plan)[keyof typeof Plan];

export const Plan = {
  starter: "starter",
  pro: "pro",
} as const;