# lib/api-zod/src/generated/types/whatsAppWebhookBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vendor Control Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface WhatsAppWebhookBody {
  /** Customer phone number */
  from: string;
  /** Vendor WhatsApp number */
  to: string;
  body: string;
  profileName?: string;
}